+++ internal/flag/flag.go
@@ -1,17 +1,29 @@
 package flag
 
-import "flag"
+import (
+	"os"
+	"strings"
+)
 
-func GetFlagValue(key, def, desc string) string {
-	var (
-		inputValue string
-	)
-
-	if flag.Lookup(key) == nil {
-		flag.StringVar(&inputValue, key, def, desc)
-		flag.Parse()
-		return inputValue
+func GetFlagValue(pKey, pDefault string) string {
+	isNextValue := false
+	for _, arg := range os.Args[1:] {
+		if isNextValue {
+			return arg
+		}
+		trimArg := strings.TrimLeft(arg, "-")
+		if !strings.HasPrefix(trimArg, pKey) {
+			continue
+		}
+		splited := strings.Split(trimArg, "=")
+		if len(splited) == 1 {
+			isNextValue = true
+			continue
+		}
+		return strings.Join(splited[1:], "=")
 	}
-
-	return flag.Lookup(key).Value.(flag.Getter).Get().(string)
+	if isNextValue {
+		panic("args has key but value is not found")
+	}
+	return pDefault
 }